refine code
