+++ paddle/fluid/CMakeLists.txt
@@ -9,7 +9,7 @@ add_subdirectory(eager)
 add_subdirectory(prim)
 add_subdirectory(jit)
 if(WITH_NEWIR)
-  add_subdirectory(paddle_dialect)
+  add_subdirectory(dialect)
 endif()
 # NOTE: please add subdirectory inference at last.
 add_subdirectory(inference)
