--- paddle/fluid/paddle_dialect/CMakeLists.txt
@@ -1,9 +1,0 @@
-set(PD_DIALECT_SOURCE_DIR "${PADDLE_SOURCE_DIR}/paddle/fluid/paddle_dialect")
-set(PD_DIALECT_BINARY_DIR "${PADDLE_BINARY_DIR}/paddle/fluid/paddle_dialect")
-
-file(GLOB PD_DIALECT_SRCS "*.cc")
-
-cc_library(
-  paddle_dialect
-  SRCS ${PD_DIALECT_SRCS}
-  DEPS new_ir framework_proto dense_tensor)
